+++ backend/storage/load-file-gcs-to-string.js
