+++ backend/storage/load-from-gcs.js
